Ignore serial noise on real HW. Fix kbd interface.
For some reason on some real hardware, the serial port continuously
returns 0xff even when idle. So ignore that in console.c.

For the kbd handling, rename kbd_getc to kdb_get_key for the function
that decodes the scan code. Create a new kbd_getc which returns the
raw scancode.